Fixed conflicts
